% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{account}
\alias{account}
\title{Get user account summary (authorization is required).}
\usage{
account(keys)
}
\arguments{
\item{keys}{list containing access and  secret keys in the form: keys <- list(key = ACCESS_KEY, secret = ACCESS_SECRET)}
}
\description{
Get user account summary (authorization is required).
}
\examples{
rderibit::account(keys)
}
