% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{cancelall}
\alias{cancelall}
\title{Bulk cancel orders based on type or instrument}
\usage{
cancelall(keys, instrument = NULL, typeDef = NULL)
}
\arguments{
\item{instrument}{string [optional], instrument name, example "BTC-28SEP18-3500-P" (or other items in rderibit::getinstruments()).}

\item{typeDef}{Cancel instructions. Can be "all", "options", "futures" or NULL. Will be ignored if instrument is not NULL}
}
\description{
Bulk cancel orders based on type or instrument
}
\examples{
cancelall(keys, "BTC-28DEC18-8750-P")
}
