% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{edit}
\alias{edit}
\title{Change order parameters}
\usage{
edit(keys, orderId, quantity, price)
}
\arguments{
\item{keys}{list containing access and  secret keys in the form: keys <- list(key = ACCESS_KEY, secret = ACCESS_SECRET)}

\item{orderId}{ID of the order to be changed}

\item{quantity}{New quantity}

\item{price}{New price}
}
\description{
Change order parameters
}
