% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{getlasttrades}
\alias{getlasttrades}
\title{Retrieve the latest trades that have occured for a specific instrument.}
\usage{
getlasttrades(instrument, count = NULL, startTimestamp = NULL,
  endTimestamp = NULL, startId = NULL, endId = NULL, startSeq = NULL,
  endSeq = NULL)
}
\arguments{
\item{instrument}{string [optional], instrument name, example "BTC-28SEP18-3500-P" (or other items in rderibit::getinstruments()). Can as well be "all", "options", "futures", "BTC-THISWEEK", "BTC-NEXTWEEK"}

\item{count}{– integer [optional], count of trades returned (limitation: max. count is 100000, default 100)}

\item{startTimestamp}{timestamp of starting trade in milliseconds}

\item{endTimestamp}{timestamp of ending trade in milliseconds}

\item{startId}{– integer [optional], “since” tradeId, the server returns trades newer than that “since”.}

\item{endId}{- integer [optional], the server returns trades with tradeId smaller than the provided number.}

\item{startSeq}{relates to tradeSeq parameter}

\item{endSeq}{relates to tradeSeq parameter}
}
\description{
Retrieve the latest trades that have occured for a specific instrument.
}
\examples{
getlasttrades(instrument = "BTC-28SEP18-3500-P", startTimestamp = as.numeric(as.POSIXlt("2018-08-05 18:00:00 UTC"))*1000)

}
