% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{orderhistory}
\alias{orderhistory}
\title{Outputs order history}
\usage{
orderhistory(keys, count = NULL, simplify = TRUE)
}
\arguments{
\item{keys}{list containing access and  secret keys in the form: keys <- list(key = ACCESS_KEY, secret = ACCESS_SECRET)}

\item{count}{(optional) Length of the output}

\item{simplify}{bool, simplifies output to dataframe, defaults to TRUE}
}
\description{
Outputs order history
}
