% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{positions}
\alias{positions}
\title{Lists open positions}
\usage{
positions(keys)
}
\arguments{
\item{keys}{list containing access and  secret keys in the form: keys <- list(key = ACCESS_KEY, secret = ACCESS_SECRET)}
}
\description{
Lists open positions
}
