% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{request}
\alias{request}
\title{Helper function to send requests to the API. If private methods are used, provide corresponding API keys}
\usage{
request(keys = NULL, action, data = NULL, simplify = TRUE)
}
\arguments{
\item{keys}{list containing access and  secret keys in the form: keys <- list(key = ACCESS_KEY, secret = ACCESS_SECRET)}

\item{action}{Action URI}

\item{data}{additional data(options) provided as list of vectors (of length one, typically)
use httr::with_verbose(rderibit::request(action = "/api/v1/private/cancel", keys = keys, data = list(orderId = "6306623820") ) ) to troubleshoot the connection to API}
}
\description{
Helper function to send requests to the API. If private methods are used, provide corresponding API keys
}
