% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deribit_api.R
\name{tradehistory}
\alias{tradehistory}
\title{Outputs trade history}
\usage{
tradehistory(keys, instrument = "all", count = NULL, startTradeId = NULL,
  simplify = TRUE)
}
\arguments{
\item{keys}{list containing access and  secret keys in the form: keys <- list(key = ACCESS_KEY, secret = ACCESS_SECRET)}

\item{instrument}{(optional) Instrument name}

\item{count}{(optional) Length of the output}

\item{startTradeId}{(optional) startTradeId where output begins}

\item{simplify}{bool, simplifies output to dataframe, defaults to TRUE}
}
\description{
Outputs trade history
}
